refactor HeroBanner by creating HeroBannerNav component; adjust button styles in categorySection for better visibility
